line_chatbot_api.py
from linebot import (
    LineBotApi, WebhookHandler
)
from linebot.exceptions import (
    InvalidSignatureError
)
from linebot.models import (
    MessageEvent, 
    PostbackEvent,
    TextMessage, 
    TextSendMessage, 
    ImageSendMessage, 
    StickerSendMessage, 
    LocationSendMessage,
    TemplateSendMessage,
    ButtonsTemplate,
    PostbackAction,
    MessageAction,
    URIAction,
    CarouselTemplate,
    CarouselColumn,
    ImageCarouselTemplate,
    ImageCarouselColumn,
    DatetimePickerAction,
    ConfirmTemplate
)

line_bot_api = LineBotApi('7JIAiRRjpm1ZoUKNVoA2w5tNxufx8QwZWX/M7kIwE71o6XaQrJdfrcSbv5mA48zyP5LXFTnJZ9sIQvABj31lBTXvt+Yd/YYjjj0g6MVbXks6t7YWpMfuzSbxPTHE13FkRrqUezfVNrgwW+EgtwA5sgdB04t89/1O/w1cDnyilFU=')
handler = WebhookHandler('9159cfc09eb833e1bdc8d4ae849010d4')

