requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.26.1
azure-storage-blob==12.14.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
Flask==2.2.2
frozenlist==1.3.3
future==0.18.2
idna==3.4
importlib-metadata==5.1.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
msrest==0.7.1
multidict==6.0.2
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
SpeechRecognition==3.8.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.11.0
